WIP better documentation and naming
