+++ src/components/Screens/EditStage.js
@@ -32,14 +32,14 @@ class EditStage extends PureComponent {
     createStage: PropTypes.func.isRequired,
     previewStage: PropTypes.func.isRequired,
     closePreview: PropTypes.func.isRequired,
-    state: PropTypes.string,
+    transitionState: PropTypes.string,
   };
 
   static defaultProps = {
     id: null,
     insertAtIndex: null,
     onComplete: () => {},
-    state: null,
+    transitionState: null,
   };
 
   get isDirty() {
@@ -88,14 +88,14 @@ class EditStage extends PureComponent {
   handleCancel = this.handleComplete;
 
   render() {
-    const { stage, show, state } = this.props;
+    const { stage, show, transitionState } = this.props;
 
     return (
       <Card
         buttons={this.buttons}
         secondaryButtons={this.secondaryButtons}
         show={show}
-        state={state}
+        transitionState={transitionState}
         onCancel={this.handleCancel}
       >
         <StageEditor
