+++ src/components/Screens/Screens.js
@@ -1,34 +1,25 @@
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { get } from 'lodash';
 import PropTypes from 'prop-types';
 import { TransitionGroup } from 'react-transition-group';
-import EditStage from './EditStage';
-import EditSkipLogic from './EditSkipLogic';
-import ViewForms from './ViewForms';
-import EditForm from './EditForm';
-import Codebook from './VariableRegistry';
-import EditType from './EditType';
-import EditVariable from './EditVariable';
 import { actionCreators as uiActions } from '../../ducks/modules/ui';
 import TimelineScreenTransition, { styles } from '../Transitions/TimelineScreen';
+import { getScreenComponent } from './screenIndex';
 
-const NotFound = () => (<div> Screen not found </div>);
-
-const NAMES = {
-  stage: EditStage,
-  skip: EditSkipLogic,
-  forms: ViewForms,
-  form: EditForm,
-  codebook: Codebook,
-  type: EditType,
-  variable: EditVariable,
-};
-
-const getScreenComponent = screen =>
-  get(NAMES, screen, NotFound);
-
+/**
+ * Screen manager for Architect.
+ *
+ * Any screens present in the `screens` redux state will be rendered by this
+ * component.
+ *
+ * Screens can be opened and closed using the screen module actions:
+ * - `openScreen(name, params)`:
+ *    `name` is used to find the screen component (screens are listed by name in screenIndex),
+ *     and `params` are passed to screen component itself
+ *    This will be rendered as `<ScreenComponent {...params} />`
+ * - `closeScreen(name, params)`
+ */
 const Screens = (props) => {
   const screens = props.screens.map(({ screen, params }, index) => {
     const ScreenComponent = getScreenComponent(screen);
@@ -41,11 +32,11 @@ const Screens = (props) => {
         style={transitionStyle}
         key={screen}
       >
-        {state => (
+        {transitionState => (
           <ScreenComponent
             {...params}
             show
-            state={state}
+            transitionState={transitionState}
             onComplete={onComplete}
           />
         )}
