+++ src/components/Screens/VariableEditorScreen.js
@@ -2,11 +2,11 @@ import { compose, withProps } from 'recompose';
 import EditScreen from './EditScreen';
 import VariableEditor, { formName } from '../VariableEditor';
 
-const EditVariable = compose(
+const VariableEditorScreen = compose(
   withProps(() => ({
     editor: VariableEditor,
     form: formName,
   })),
 )(EditScreen);
 
-export default EditVariable;
+export default VariableEditorScreen;
